script/requirements.txt
ecdsa==0.13
requests==2.18.4
APScheduler==3.5.1
logzero==1.3.1
base58==0.2.5
bitcoin==1.1.42
pycoin==0.80
pymongo==3.6.0
python-dotenv==0.8.0
redis==2.10.6
pycryptodomex==3.6.5
cryptography==2.3.1